% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prj.R
\name{create_prj}
\alias{create_prj}
\title{create_prj}
\usage{
create_prj(db_name, base_path, desired_scen = NULL, prj_name = NULL)
}
\arguments{
\item{db_name}{name of the database. It will The extension will be automatically added if not present}

\item{base_path}{run directory containing the `output/` (GCAM databases)
and `prj_files/` folders. See gcam_sdg_base_path().}

\item{desired_scen}{desired scenarios. If NULL, all the scenarios present in the database will be considered}

\item{prj_name}{name of the project. If NULL, it will be the defult option, i.e., the database name. Otherwise specify}
}
\value{
create the specified project
}
\description{
Function to create a GCAM project provided a database and a queries file
}
