% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDG_indicators.R
\name{extract_data}
\alias{extract_data}
\title{extract_data}
\usage{
extract_data(dat_list, pre_path)
}
\arguments{
\item{dat_list}{file names (relative to pre_path) to read and bind}

\item{pre_path}{directory containing the files in dat_list}
}
\value{
combined data frame
}
\description{
Read and bind a list of gathered SDG indicator .RData/.csv files into a
single data frame.
}
