% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg15_land_indicator.R
\name{get_sdg15_land_indicator}
\alias{get_sdg15_land_indicator}
\title{get_sdg15_land_indicator}
\usage{
get_sdg15_land_indicator(prj, prj_name, saveOutput = T, makeFigures = F)
}
\arguments{
\item{prj}{uploaded project file}

\item{prj_name}{project file name, used to tag the saved output file}

\item{saveOutput}{save the produced output}

\item{makeFigures}{generate and save graphical representation/s of the output}
}
\value{
data frame with the final PSL by scenario
}
\description{
Compute SDG 15 (Life on Land) as the net Potential Species Loss (PSL)
indicator, downscaling GCAM land allocation with Demeter and aggregating
land-use change to the ecoregion level.
}
