% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDG_indicators.R
\name{run_comparisson}
\alias{run_comparisson}
\title{run_comparisson}
\usage{
run_comparisson(ssp, base_path, final_db_year = 2050)
}
\arguments{
\item{ssp}{SSP tag to filter scenarios by}

\item{base_path}{directory containing the gathered SDGx-... indicator
subfolders (the same value used when the indicators were gathered)}

\item{final_db_year}{last model year to consider}
}
\value{
list of data frames: gdp, expenditure, poverty, health, water, land
}
\description{
Compute the final SDG indicators (SSP vs. REF) from the gathered
per-scenario indicator files written by gather_indicators().
}
